ai/apps/DG_collect_dataset/requirements.txt
# DG_collect_dataset - Dataset Collection Pipeline Requirements
# This file lists all dependencies for the complete pipeline

# === Core ML/AI ===

# === Core ML/AI ===
torch>=2.0.0
transformers>=4.45.0
accelerate>=0.25.0
bitsandbytes>=0.42.0
safetensors>=0.4.0
qwen-vl-utils>=0.0.14
huggingface-hub>=0.20.0

# === Image Processing ===
torch>=2.0.0
transformers>=4.45.0
accelerate>=0.25.0
bitsandbytes>=0.42.0
safetensors>=0.4.0
qwen-vl-utils>=0.0.14
huggingface-hub>=0.20.0
Pillow>=10.0.0
opencv-python>=4.8.0
insightface>=0.7.3
tensorflow>=2.15.0
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
cryptography>=41.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
PyYAML>=6.0.0
PyYAML>=6.0.0
